Moving zip utilities in the archives module
